refactor: implement Blueprints
This may or may not work and will require more testing.
