--- run.py
@@ -1,4 +1,0 @@
-from __init__ import create_app
-
-app = create_app()
-# run app here
